Fix Extract Notebook and Module (+1)
Also change .gitignore to ignore checkpoints.

+++ captioner/extract.py
@@ -49,17 +49,18 @@ def __main():
 	from pathlib import Path
 	from hparams import image_shape, architecture, num_workers
 	from hparams import extractor_batch_size as batch_size
+	from captioner.data import get_extract_dataloaders
 
 	DIR_DATA = Path('~/.data/COCO').expanduser()
 
-	#if (DIR_DATA / 'train/features.pt').exists(): return
+	if (DIR_DATA / 'train/features.pt').exists(): return
 
-	dataloader = get_dataloaders(DIR_DATA, image_shape, batch_size, num_workers)
+	dataloader = get_extract_dataloaders(DIR_DATA, image_shape, batch_size, num_workers)
 	extractor = Extractor(architecture)
 
-	for mode in ('train', ):
+	for mode in ('val', 'train'):
 		print(f'Extracting features for set {mode}')
-		features = extract(extractor, dataloader[mode])
-		torch.save(features.to('cpu'), DIR_DATA / mode / 'features.pt')
+		with mag.eval(extractor): features = extractor(dataloader[mode])
+    	torch.save(features.to('cpu'), DIR_DATA / mode / 'features.pt')
 
 if __name__ == '__main__': __main()
